Fix bug in PIDController, add helpful debug messages.

+++ src/PIDController.h
@@ -12,17 +12,17 @@
 class PIDController
 {
 public:
-  PIDController(float target, float Kp, float Ki, float Kd, float dropoff);
+  PIDController(double target, double Kp, double Ki, double Kd, double dropoff);
   byte add_reading(SensorReading reading, time_t timestamp);
 
 private:
-  float _target = 0;
-  float _dropoff = 0;
-  float _Kp = 0;
-  float _Ki = 0;
-  float _Kd = 0;
-  float _prev_error = 0;
-  float _integral = 0;
+  double _target = 0;
+  double _dropoff = 0;
+  double _Kp = 0;
+  double _Ki = 0;
+  double _Kd = 0;
+  double _prev_error = 0;
+  double _integral = 0;
   time_t last_sampled = 0;
 };
